+++ src/array/array_sync_writable.rs
@@ -313,11 +313,11 @@ impl<TStorage: ?Sized + WritableStorageTraits> Array<TStorage> {
 
     /// Erase the chunk at `chunk_indices`.
     ///
-    /// Returns true if the chunk was erased, or false if it did not exist.
+    /// Succeeds if the chunk does not exist.
     ///
     /// # Errors
     /// Returns a [`StorageError`] if there is an underlying store error.
-    pub fn erase_chunk(&self, chunk_indices: &[u64]) -> Result<bool, StorageError> {
+    pub fn erase_chunk(&self, chunk_indices: &[u64]) -> Result<(), StorageError> {
         let storage_handle = Arc::new(StorageHandle::new(&*self.storage));
         let storage_transformer = self
             .storage_transformers()