+++ src/storage/storage_transformer/performance_metrics.rs
@@ -269,15 +269,15 @@ impl<TStorage: ?Sized + WritableStorageTraits> WritableStorageTraits
         self.storage.set_partial_values(key_start_values)
     }
 
-    fn erase(&self, key: &StoreKey) -> Result<bool, StorageError> {
+    fn erase(&self, key: &StoreKey) -> Result<(), StorageError> {
         self.storage.erase(key)
     }
 
-    fn erase_values(&self, keys: &[StoreKey]) -> Result<bool, StorageError> {
+    fn erase_values(&self, keys: &[StoreKey]) -> Result<(), StorageError> {
         self.storage.erase_values(keys)
     }
 
-    fn erase_prefix(&self, prefix: &StorePrefix) -> Result<bool, StorageError> {
+    fn erase_prefix(&self, prefix: &StorePrefix) -> Result<(), StorageError> {
         self.storage.erase_prefix(prefix)
     }
 }
@@ -408,15 +408,15 @@ impl<TStorage: ?Sized + AsyncWritableStorageTraits> AsyncWritableStorageTraits
         self.storage.set_partial_values(key_start_values).await
     }
 
-    async fn erase(&self, key: &StoreKey) -> Result<bool, StorageError> {
+    async fn erase(&self, key: &StoreKey) -> Result<(), StorageError> {
         self.storage.erase(key).await
     }
 
-    async fn erase_values(&self, keys: &[StoreKey]) -> Result<bool, StorageError> {
+    async fn erase_values(&self, keys: &[StoreKey]) -> Result<(), StorageError> {
         self.storage.erase_values(keys).await
     }
 
-    async fn erase_prefix(&self, prefix: &StorePrefix) -> Result<bool, StorageError> {
+    async fn erase_prefix(&self, prefix: &StorePrefix) -> Result<(), StorageError> {
         self.storage.erase_prefix(prefix).await
     }
 }