+++ src/storage/store/store_async.rs
@@ -46,6 +46,7 @@ mod test_util {
 
         store.set(&"erase".try_into()?, &[]).await?;
         store.erase(&"erase".try_into()?).await?;
+        store.erase(&"erase".try_into()?).await?; // succeeds
 
         store.set(&"erase_values_0".try_into()?, &[]).await?;
         store.set(&"erase_values_1".try_into()?, &[]).await?;